Group layer (#50)
* Added groups to database layer. Creating links right now between user and groups, for ease of updation in case user details change

* Added groups to database layer. Creating links right now between user and groups, for ease of updation in case user details change

* set up a preliminary version of create_group with User holding links to Group

* added privilege functionalities to groups, need to workout better way of implementing the privileges with mongo db friendly enums

* Updated group access_controls to Enum types, needed to define Accesscontrol(str, Enum)

* Fixed groups field in User Document

* Added update feature onto user groups field to update existing groups field in users

* Added add_group/remove_group feature in user update service

* cleaning database definitions;updating group service to use lists similar to Users

* reverting test_product_service.py to original state

* Updated function names and cleaned unnecessary error handling at group creation

* Added independent update function for modifying access controls

* Removed possibility of access_control to be None and updated test service for groups to check for id

* Added a group api layer, test layer and updated auth

* added update_group_access in api layer

* Added update description and delete group api endpoint

* updated users api and service layer for update_groups

* Added feature for creating the common users group in the group service layer

* Moved group privilege check to service

* moved privilege check to group level from api layer to service layer

* moved common group creation to group service layer

* updated error in parameter definition

* Updated database defining ProtectedDocument which is inherited by Product

* Introducing readers and writers to the product

* Check for groups when accessing products

* modified test and update feature in product service layer

* Updated Product layer in service and resolved merge conflicts

* Updating API layers to take into account changes in product services

* Additional updates to service layer of products and translated updates acrosss api layer

* Added Api layer updates to products for group records

* Updated collection to be preotected document

* Translated updates of protected document to relationships api

* Added authentication checks in collection service, some product service updates

* applying product filtration at collection level

* catching product http exception during filtration

* added readers and writers to collection api

* Added cli features to collections and products

+++ hipposerve/service/collection.py
@@ -7,8 +7,14 @@ the product service.
 
 from beanie import PydanticObjectId
 from beanie.operators import Text
+from fastapi import HTTPException
 
-from hipposerve.database import Collection
+from hipposerve.database import Collection, User
+from hipposerve.service.auth import (
+    check_collection_read_access,
+    check_collection_write_access,
+    check_product_read_access,
+)
 
 
 class CollectionNotFound(Exception):
@@ -17,13 +23,24 @@ class CollectionNotFound(Exception):
 
 async def create(
     name: str,
+    user: User,
     description: str,
+    collection_readers: list[str] | None = None,
+    collection_writers: list[str] | None = None,
 ):
     collection = Collection(
         name=name,
+        owner=user,
         description=description,
     )
+    collection_readers = collection_readers or []
+    collection_writers = collection_writers or []
 
+    if collection_readers is not None:
+        collection.readers.extend(collection_readers)
+    if user.name not in collection.writers:
+        collection_writers.append(user.name)
+    collection.writers.extend(collection_writers)
     await collection.insert()
 
     return collection
@@ -31,42 +48,87 @@ async def create(
 
 async def read(
     id: PydanticObjectId,
+    user: User,
 ):
     collection = await Collection.find_one(Collection.id == id, fetch_links=True)
 
     if collection is None:
         raise CollectionNotFound
-
+    await check_collection_read_access(user, collection)
+    if collection.products:
+        collection_list = await collection_product_filter(user, [collection])
+        collection = collection_list[0]
     return collection
 
 
 async def read_most_recent(
-    fetch_links: bool = False, maximum: int = 16
+    user: User,
+    fetch_links: bool = False,
+    maximum: int = 16,
 ) -> list[Collection]:
     # TODO: Implement updated time for collections.
-    return await Collection.find(fetch_links=fetch_links).to_list(maximum)
+    group_names = [group.name for group in user.groups]
+    access_query = {
+        "$or": [{"readers": {"$in": group_names}}, {"writers": {"$in": group_names}}]
+    }
+    collection_list = await Collection.find(
+        access_query, fetch_links=fetch_links
+    ).to_list(maximum)
+    filtered_collection_list = await collection_product_filter(user, collection_list)
+    return filtered_collection_list
 
 
-async def search_by_name(name: str, fetch_links: bool = True) -> list[Collection]:
+async def search_by_name(
+    name: str, user: User, fetch_links: bool = True
+) -> list[Collection]:
     """
     Search for Collections by name using the text index.
     """
-
+    group_names = [group.name for group in user.groups]
+    access_query = {
+        "$or": [{"readers": {"$in": group_names}}, {"writers": {"$in": group_names}}]
+    }
     results = (
-        await Collection.find(Text(name), fetch_links=fetch_links)  # noqa: E712
+        await Collection.find(access_query, Text(name), fetch_links=fetch_links)  # noqa: E712
         .sort([("score", {"$meta": "textScore"})])
         .to_list()
     )
 
-    return results
+    filtered_collection_list = await collection_product_filter(user, results)
+    return filtered_collection_list
 
 
 async def update(
     id: PydanticObjectId,
+    access_user: User,
+    owner: User | None,
     description: str | None,
+    add_readers: list[str] | None = None,
+    remove_readers: list[str] | None = None,
+    add_writers: list[str] | None = None,
+    remove_writers: list[str] | None = None,
 ):
-    collection = await read(id=id)
-
+    collection = await read(id=id, user=access_user)
+    await check_collection_write_access(access_user, collection)
+    readers = collection.readers.copy()
+    writers = collection.writers.copy()
+    # We don't actually 'update' the database; we actually create a new
+    # product and link it in.
+    if owner is not None:
+        await collection.set({Collection.owner: owner})
+        writers.append(owner.name)
+    if add_readers is not None:
+        readers.extend(add_readers)
+        await collection.set({Collection.readers: readers})
+    if remove_readers is not None:
+        readers = [x for x in readers if x not in remove_readers]
+        await collection.set({Collection.readers: readers})
+    if add_writers is not None:
+        writers.extend(add_writers)
+        await collection.set({Collection.writers: writers})
+    if remove_writers is not None:
+        writers = [x for x in writers if x not in remove_writers]
+        await collection.set({Collection.writers: writers})
     if description is not None:
         await collection.set({Collection.description: description})
 
@@ -76,10 +138,11 @@ async def update(
 async def add_child(
     parent_id: PydanticObjectId,
     child_id: PydanticObjectId,
+    user: User,
 ) -> Collection:
-    parent = await read(id=parent_id)
-    child = await read(id=child_id)
-
+    parent = await read(id=parent_id, user=user)
+    child = await read(id=child_id, user=user)
+    await check_collection_write_access(user, parent)
     parent.child_collections.append(child)
     await parent.save()
 
@@ -89,9 +152,10 @@ async def add_child(
 async def remove_child(
     parent_id: PydanticObjectId,
     child_id: PydanticObjectId,
+    user: User,
 ) -> Collection:
-    parent = await read(id=parent_id)
-
+    parent = await read(id=parent_id, user=user)
+    await check_collection_write_access(user, parent)
     await parent.set(
         {
             Collection.child_collections: [
@@ -105,9 +169,29 @@ async def remove_child(
 
 async def delete(
     id: PydanticObjectId,
+    user: User,
 ):
-    collection = await read(id=id)
-
+    collection = await read(id=id, user=user)
+    await check_collection_write_access(user, collection)
     await collection.delete()
 
     return
+
+
+async def collection_product_filter(
+    user: User,
+    collection_list: list[Collection],
+) -> list[Collection]:
+    filtered_collection_list = []
+    for collection in collection_list:
+        if collection.products:
+            product_list = []
+            for product in collection.products:
+                try:
+                    await check_product_read_access(user, product)
+                    product_list.append(product)
+                except HTTPException:
+                    continue
+            collection.products = product_list
+        filtered_collection_list.append(collection)
+    return filtered_collection_list
